+++ database/mysql/inventario_base.sql
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS productos_granel (
 CREATE TABLE IF NOT EXISTS ventas_historial (
     id INT AUTO_INCREMENT PRIMARY KEY,
     tipo_pago ENUM('efectivo', 'tarjeta') NOT NULL,
-    total DECIMAL(10,2) NOT NULL,
+    total INT NOT NULL,
     fecha_venta DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
@@ -50,9 +50,9 @@ CREATE TABLE IF NOT EXISTS ventas_historial_detalle (
     producto_id INT NOT NULL,
     codigo_producto VARCHAR(50) NOT NULL,
     nombre_producto VARCHAR(150) NOT NULL,
-    cantidad DECIMAL(10,2) NOT NULL,
+    cantidad INT NOT NULL,
     unidad_medida VARCHAR(20) NOT NULL,
-    precio_referencia DECIMAL(10,2) NOT NULL,
-    subtotal DECIMAL(10,2) NOT NULL,
+    precio_referencia INT NOT NULL,
+    subtotal INT NOT NULL,
     FOREIGN KEY (venta_id) REFERENCES ventas_historial(id)
 );
